added dytpe=float64 to all other pd.Series() as well

+++ pandapipes/test/openmodelica_comparison/pipeflow_openmodelica_comparison.py
@@ -116,8 +116,8 @@ def pipeflow_openmodelica_comparison(net, log_results=True, friction_model='cole
     p_pandapipes = net.res_junction.p_bar.loc[p_valid].values.astype(np.float64).round(4)
     p_diff = np.abs(1 - p_pandapipes / p_om)
 
-    p_diff = pd.Series(p_diff, range(len(p_diff)))
-    v_diff_abs = pd.Series(v_diff_abs, range(len(v_diff_abs)))
+    p_diff = pd.Series(p_diff, range(len(p_diff)), dtype='float64')
+    v_diff_abs = pd.Series(v_diff_abs, range(len(v_diff_abs)), dtype='float64')
 
     '''
     print("\n p_diff = \n", p_diff)
@@ -178,9 +178,9 @@ def retrieve_velocity_liquid(net, element="pipe"):
     v_diff_mean = np.abs(1 - v_mean_pandapipes / v_om)
     v_diff_abs = np.abs(v_om - v_mean_pandapipes)
 
-    v_diff_mean = pd.Series(v_diff_mean, range(len(v_diff_mean)))
-    v_diff_abs = pd.Series(v_diff_abs, range(len(v_diff_abs)))
-    v_om = pd.Series(v_om, range(len(v_om)))
+    v_diff_mean = pd.Series(v_diff_mean, range(len(v_diff_mean)), dtype='float64')
+    v_diff_abs = pd.Series(v_diff_abs, range(len(v_diff_abs)), dtype='float64')
+    v_om = pd.Series(v_om, range(len(v_om)), dtype='float64')
 
     return v_diff_mean, v_diff_abs, v_mean_pandapipes, v_om
 
@@ -219,11 +219,11 @@ def retrieve_velocity_gas(net, element='pipe'):
     v_diff_mean = np.abs(1 - v_mean_pandapipes / v_om)
     v_diff_abs = np.abs(v_om - v_mean_pandapipes)
 
-    v_diff_mean = pd.Series(v_diff_mean, range(len(v_diff_mean)))
-    v_diff_from = pd.Series(v_diff_from, range(len(v_diff_from)))
-    v_diff_to = pd.Series(v_diff_to, range(len(v_diff_to)))
-    v_diff_abs = pd.Series(v_diff_abs, range(len(v_diff_abs)))
-    v_om = pd.Series(v_om, range(len(v_om)))
+    v_diff_mean = pd.Series(v_diff_mean, range(len(v_diff_mean)), dtype='float64')
+    v_diff_from = pd.Series(v_diff_from, range(len(v_diff_from)), dtype='float64')
+    v_diff_to = pd.Series(v_diff_to, range(len(v_diff_to)), dtype='float64')
+    v_diff_abs = pd.Series(v_diff_abs, range(len(v_diff_abs)), dtype='float64')
+    v_om = pd.Series(v_om, range(len(v_om)), dtype='float64')
 
     return v_diff_from, v_diff_to, v_diff_mean, v_diff_abs, v_mean_pandapipes, v_om
 
@@ -254,10 +254,10 @@ def retrieve_temperature_liquid(net):
     T_diff_mean = np.abs(1 - T_mean_pandapipes / T_mean_om)
     T_diff_abs = np.abs(T_mean_om - T_mean_pandapipes)
 
-    T_diff_mean = pd.Series(T_diff_mean, range(len(T_diff_mean)))
-    T_diff_abs = pd.Series(T_diff_abs, range(len(T_diff_abs)))
+    T_diff_mean = pd.Series(T_diff_mean, range(len(T_diff_mean)), dtype='float64')
+    T_diff_abs = pd.Series(T_diff_abs, range(len(T_diff_abs)), dtype='float64')
 
-    T_om = pd.Series(T_om, range(len(T_om)))
-    T_mean_pandapipes = pd.Series(T_mean_pandapipes, range(len(T_mean_pandapipes)))
+    T_om = pd.Series(T_om, range(len(T_om)), dtype='float64')
+    T_mean_pandapipes = pd.Series(T_mean_pandapipes, range(len(T_mean_pandapipes)), dtype='float64')
 
     return T_diff_mean, T_diff_abs, T_mean_pandapipes, T_om
